moving those files weirded it out
