+++ lib/tree/ast/antlr/java/antlr-gen/Java9Visitor.ts
@@ -3,251 +3,251 @@
 
 import { ParseTreeVisitor } from "antlr4ts/tree/ParseTreeVisitor";
 
-import { LiteralContext } from "../Java9Parser";
-import { PrimitiveTypeContext } from "../Java9Parser";
-import { NumericTypeContext } from "../Java9Parser";
-import { IntegralTypeContext } from "../Java9Parser";
-import { FloatingPointTypeContext } from "../Java9Parser";
-import { ReferenceTypeContext } from "../Java9Parser";
-import { ClassOrInterfaceTypeContext } from "../Java9Parser";
-import { ClassTypeContext } from "../Java9Parser";
-import { ClassType_lf_classOrInterfaceTypeContext } from "../Java9Parser";
-import { ClassType_lfno_classOrInterfaceTypeContext } from "../Java9Parser";
-import { InterfaceTypeContext } from "../Java9Parser";
-import { InterfaceType_lf_classOrInterfaceTypeContext } from "../Java9Parser";
-import { InterfaceType_lfno_classOrInterfaceTypeContext } from "../Java9Parser";
-import { TypeVariableContext } from "../Java9Parser";
-import { ArrayTypeContext } from "../Java9Parser";
-import { DimsContext } from "../Java9Parser";
-import { TypeParameterContext } from "../Java9Parser";
-import { TypeParameterModifierContext } from "../Java9Parser";
-import { TypeBoundContext } from "../Java9Parser";
-import { AdditionalBoundContext } from "../Java9Parser";
-import { TypeArgumentsContext } from "../Java9Parser";
-import { TypeArgumentListContext } from "../Java9Parser";
-import { TypeArgumentContext } from "../Java9Parser";
-import { WildcardContext } from "../Java9Parser";
-import { WildcardBoundsContext } from "../Java9Parser";
-import { ModuleNameContext } from "../Java9Parser";
-import { PackageNameContext } from "../Java9Parser";
-import { TypeNameContext } from "../Java9Parser";
-import { PackageOrTypeNameContext } from "../Java9Parser";
-import { ExpressionNameContext } from "../Java9Parser";
-import { MethodNameContext } from "../Java9Parser";
-import { AmbiguousNameContext } from "../Java9Parser";
-import { CompilationUnitContext } from "../Java9Parser";
-import { OrdinaryCompilationContext } from "../Java9Parser";
-import { ModularCompilationContext } from "../Java9Parser";
-import { PackageDeclarationContext } from "../Java9Parser";
-import { PackageModifierContext } from "../Java9Parser";
-import { ImportDeclarationContext } from "../Java9Parser";
-import { SingleTypeImportDeclarationContext } from "../Java9Parser";
-import { TypeImportOnDemandDeclarationContext } from "../Java9Parser";
-import { SingleStaticImportDeclarationContext } from "../Java9Parser";
-import { StaticImportOnDemandDeclarationContext } from "../Java9Parser";
-import { TypeDeclarationContext } from "../Java9Parser";
-import { ModuleDeclarationContext } from "../Java9Parser";
-import { ModuleDirectiveContext } from "../Java9Parser";
-import { RequiresModifierContext } from "../Java9Parser";
-import { ClassDeclarationContext } from "../Java9Parser";
-import { NormalClassDeclarationContext } from "../Java9Parser";
-import { ClassModifierContext } from "../Java9Parser";
-import { TypeParametersContext } from "../Java9Parser";
-import { TypeParameterListContext } from "../Java9Parser";
-import { SuperclassContext } from "../Java9Parser";
-import { SuperinterfacesContext } from "../Java9Parser";
-import { InterfaceTypeListContext } from "../Java9Parser";
-import { ClassBodyContext } from "../Java9Parser";
-import { ClassBodyDeclarationContext } from "../Java9Parser";
-import { ClassMemberDeclarationContext } from "../Java9Parser";
-import { FieldDeclarationContext } from "../Java9Parser";
-import { FieldModifierContext } from "../Java9Parser";
-import { VariableDeclaratorListContext } from "../Java9Parser";
-import { VariableDeclaratorContext } from "../Java9Parser";
-import { VariableDeclaratorIdContext } from "../Java9Parser";
-import { VariableInitializerContext } from "../Java9Parser";
-import { UnannTypeContext } from "../Java9Parser";
-import { UnannPrimitiveTypeContext } from "../Java9Parser";
-import { UnannReferenceTypeContext } from "../Java9Parser";
-import { UnannClassOrInterfaceTypeContext } from "../Java9Parser";
-import { UnannClassTypeContext } from "../Java9Parser";
-import { UnannClassType_lf_unannClassOrInterfaceTypeContext } from "../Java9Parser";
-import { UnannClassType_lfno_unannClassOrInterfaceTypeContext } from "../Java9Parser";
-import { UnannInterfaceTypeContext } from "../Java9Parser";
-import { UnannInterfaceType_lf_unannClassOrInterfaceTypeContext } from "../Java9Parser";
-import { UnannInterfaceType_lfno_unannClassOrInterfaceTypeContext } from "../Java9Parser";
-import { UnannTypeVariableContext } from "../Java9Parser";
-import { UnannArrayTypeContext } from "../Java9Parser";
-import { MethodDeclarationContext } from "../Java9Parser";
-import { MethodModifierContext } from "../Java9Parser";
-import { MethodHeaderContext } from "../Java9Parser";
-import { ResultContext } from "../Java9Parser";
-import { MethodDeclaratorContext } from "../Java9Parser";
-import { FormalParameterListContext } from "../Java9Parser";
-import { FormalParametersContext } from "../Java9Parser";
-import { FormalParameterContext } from "../Java9Parser";
-import { VariableModifierContext } from "../Java9Parser";
-import { LastFormalParameterContext } from "../Java9Parser";
-import { ReceiverParameterContext } from "../Java9Parser";
-import { Throws_Context } from "../Java9Parser";
-import { ExceptionTypeListContext } from "../Java9Parser";
-import { ExceptionTypeContext } from "../Java9Parser";
-import { MethodBodyContext } from "../Java9Parser";
-import { InstanceInitializerContext } from "../Java9Parser";
-import { StaticInitializerContext } from "../Java9Parser";
-import { ConstructorDeclarationContext } from "../Java9Parser";
-import { ConstructorModifierContext } from "../Java9Parser";
-import { ConstructorDeclaratorContext } from "../Java9Parser";
-import { SimpleTypeNameContext } from "../Java9Parser";
-import { ConstructorBodyContext } from "../Java9Parser";
-import { ExplicitConstructorInvocationContext } from "../Java9Parser";
-import { EnumDeclarationContext } from "../Java9Parser";
-import { EnumBodyContext } from "../Java9Parser";
-import { EnumConstantListContext } from "../Java9Parser";
-import { EnumConstantContext } from "../Java9Parser";
-import { EnumConstantModifierContext } from "../Java9Parser";
-import { EnumBodyDeclarationsContext } from "../Java9Parser";
-import { InterfaceDeclarationContext } from "../Java9Parser";
-import { NormalInterfaceDeclarationContext } from "../Java9Parser";
-import { InterfaceModifierContext } from "../Java9Parser";
-import { ExtendsInterfacesContext } from "../Java9Parser";
-import { InterfaceBodyContext } from "../Java9Parser";
-import { InterfaceMemberDeclarationContext } from "../Java9Parser";
-import { ConstantDeclarationContext } from "../Java9Parser";
-import { ConstantModifierContext } from "../Java9Parser";
-import { InterfaceMethodDeclarationContext } from "../Java9Parser";
-import { InterfaceMethodModifierContext } from "../Java9Parser";
-import { AnnotationTypeDeclarationContext } from "../Java9Parser";
-import { AnnotationTypeBodyContext } from "../Java9Parser";
-import { AnnotationTypeMemberDeclarationContext } from "../Java9Parser";
-import { AnnotationTypeElementDeclarationContext } from "../Java9Parser";
-import { AnnotationTypeElementModifierContext } from "../Java9Parser";
-import { DefaultValueContext } from "../Java9Parser";
-import { AnnotationContext } from "../Java9Parser";
-import { NormalAnnotationContext } from "../Java9Parser";
-import { ElementValuePairListContext } from "../Java9Parser";
-import { ElementValuePairContext } from "../Java9Parser";
-import { ElementValueContext } from "../Java9Parser";
-import { ElementValueArrayInitializerContext } from "../Java9Parser";
-import { ElementValueListContext } from "../Java9Parser";
-import { MarkerAnnotationContext } from "../Java9Parser";
-import { SingleElementAnnotationContext } from "../Java9Parser";
-import { ArrayInitializerContext } from "../Java9Parser";
-import { VariableInitializerListContext } from "../Java9Parser";
-import { BlockContext } from "../Java9Parser";
-import { BlockStatementsContext } from "../Java9Parser";
-import { BlockStatementContext } from "../Java9Parser";
-import { LocalVariableDeclarationStatementContext } from "../Java9Parser";
-import { LocalVariableDeclarationContext } from "../Java9Parser";
-import { StatementContext } from "../Java9Parser";
-import { StatementNoShortIfContext } from "../Java9Parser";
-import { StatementWithoutTrailingSubstatementContext } from "../Java9Parser";
-import { EmptyStatementContext } from "../Java9Parser";
-import { LabeledStatementContext } from "../Java9Parser";
-import { LabeledStatementNoShortIfContext } from "../Java9Parser";
-import { ExpressionStatementContext } from "../Java9Parser";
-import { StatementExpressionContext } from "../Java9Parser";
-import { IfThenStatementContext } from "../Java9Parser";
-import { IfThenElseStatementContext } from "../Java9Parser";
-import { IfThenElseStatementNoShortIfContext } from "../Java9Parser";
-import { AssertStatementContext } from "../Java9Parser";
-import { SwitchStatementContext } from "../Java9Parser";
-import { SwitchBlockContext } from "../Java9Parser";
-import { SwitchBlockStatementGroupContext } from "../Java9Parser";
-import { SwitchLabelsContext } from "../Java9Parser";
-import { SwitchLabelContext } from "../Java9Parser";
-import { EnumConstantNameContext } from "../Java9Parser";
-import { WhileStatementContext } from "../Java9Parser";
-import { WhileStatementNoShortIfContext } from "../Java9Parser";
-import { DoStatementContext } from "../Java9Parser";
-import { ForStatementContext } from "../Java9Parser";
-import { ForStatementNoShortIfContext } from "../Java9Parser";
-import { BasicForStatementContext } from "../Java9Parser";
-import { BasicForStatementNoShortIfContext } from "../Java9Parser";
-import { ForInitContext } from "../Java9Parser";
-import { ForUpdateContext } from "../Java9Parser";
-import { StatementExpressionListContext } from "../Java9Parser";
-import { EnhancedForStatementContext } from "../Java9Parser";
-import { EnhancedForStatementNoShortIfContext } from "../Java9Parser";
-import { BreakStatementContext } from "../Java9Parser";
-import { ContinueStatementContext } from "../Java9Parser";
-import { ReturnStatementContext } from "../Java9Parser";
-import { ThrowStatementContext } from "../Java9Parser";
-import { SynchronizedStatementContext } from "../Java9Parser";
-import { TryStatementContext } from "../Java9Parser";
-import { CatchesContext } from "../Java9Parser";
-import { CatchClauseContext } from "../Java9Parser";
-import { CatchFormalParameterContext } from "../Java9Parser";
-import { CatchTypeContext } from "../Java9Parser";
-import { Finally_Context } from "../Java9Parser";
-import { TryWithResourcesStatementContext } from "../Java9Parser";
-import { ResourceSpecificationContext } from "../Java9Parser";
-import { ResourceListContext } from "../Java9Parser";
-import { ResourceContext } from "../Java9Parser";
-import { VariableAccessContext } from "../Java9Parser";
-import { PrimaryContext } from "../Java9Parser";
-import { PrimaryNoNewArrayContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lf_arrayAccessContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lfno_arrayAccessContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lf_primaryContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lf_primary_lf_arrayAccess_lf_primaryContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lf_primary_lfno_arrayAccess_lf_primaryContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lfno_primaryContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lfno_primary_lf_arrayAccess_lfno_primaryContext } from "../Java9Parser";
-import { PrimaryNoNewArray_lfno_primary_lfno_arrayAccess_lfno_primaryContext } from "../Java9Parser";
-import { ClassLiteralContext } from "../Java9Parser";
-import { ClassInstanceCreationExpressionContext } from "../Java9Parser";
-import { ClassInstanceCreationExpression_lf_primaryContext } from "../Java9Parser";
-import { ClassInstanceCreationExpression_lfno_primaryContext } from "../Java9Parser";
-import { TypeArgumentsOrDiamondContext } from "../Java9Parser";
-import { FieldAccessContext } from "../Java9Parser";
-import { FieldAccess_lf_primaryContext } from "../Java9Parser";
-import { FieldAccess_lfno_primaryContext } from "../Java9Parser";
-import { ArrayAccessContext } from "../Java9Parser";
-import { ArrayAccess_lf_primaryContext } from "../Java9Parser";
-import { ArrayAccess_lfno_primaryContext } from "../Java9Parser";
-import { MethodInvocationContext } from "../Java9Parser";
-import { MethodInvocation_lf_primaryContext } from "../Java9Parser";
-import { MethodInvocation_lfno_primaryContext } from "../Java9Parser";
-import { ArgumentListContext } from "../Java9Parser";
-import { MethodReferenceContext } from "../Java9Parser";
-import { MethodReference_lf_primaryContext } from "../Java9Parser";
-import { MethodReference_lfno_primaryContext } from "../Java9Parser";
-import { ArrayCreationExpressionContext } from "../Java9Parser";
-import { DimExprsContext } from "../Java9Parser";
-import { DimExprContext } from "../Java9Parser";
-import { ConstantExpressionContext } from "../Java9Parser";
-import { ExpressionContext } from "../Java9Parser";
-import { LambdaExpressionContext } from "../Java9Parser";
-import { LambdaParametersContext } from "../Java9Parser";
-import { InferredFormalParameterListContext } from "../Java9Parser";
-import { LambdaBodyContext } from "../Java9Parser";
-import { AssignmentExpressionContext } from "../Java9Parser";
-import { AssignmentContext } from "../Java9Parser";
-import { LeftHandSideContext } from "../Java9Parser";
-import { AssignmentOperatorContext } from "../Java9Parser";
-import { ConditionalExpressionContext } from "../Java9Parser";
-import { ConditionalOrExpressionContext } from "../Java9Parser";
-import { ConditionalAndExpressionContext } from "../Java9Parser";
-import { InclusiveOrExpressionContext } from "../Java9Parser";
-import { ExclusiveOrExpressionContext } from "../Java9Parser";
-import { AndExpressionContext } from "../Java9Parser";
-import { EqualityExpressionContext } from "../Java9Parser";
-import { RelationalExpressionContext } from "../Java9Parser";
-import { ShiftExpressionContext } from "../Java9Parser";
-import { AdditiveExpressionContext } from "../Java9Parser";
-import { MultiplicativeExpressionContext } from "../Java9Parser";
-import { UnaryExpressionContext } from "../Java9Parser";
-import { PreIncrementExpressionContext } from "../Java9Parser";
-import { PreDecrementExpressionContext } from "../Java9Parser";
-import { UnaryExpressionNotPlusMinusContext } from "../Java9Parser";
-import { PostfixExpressionContext } from "../Java9Parser";
-import { PostIncrementExpressionContext } from "../Java9Parser";
-import { PostIncrementExpression_lf_postfixExpressionContext } from "../Java9Parser";
-import { PostDecrementExpressionContext } from "../Java9Parser";
-import { PostDecrementExpression_lf_postfixExpressionContext } from "../Java9Parser";
-import { CastExpressionContext } from "../Java9Parser";
-import { IdentifierContext } from "../Java9Parser";
+import { LiteralContext } from "./Java9Parser";
+import { PrimitiveTypeContext } from "./Java9Parser";
+import { NumericTypeContext } from "./Java9Parser";
+import { IntegralTypeContext } from "./Java9Parser";
+import { FloatingPointTypeContext } from "./Java9Parser";
+import { ReferenceTypeContext } from "./Java9Parser";
+import { ClassOrInterfaceTypeContext } from "./Java9Parser";
+import { ClassTypeContext } from "./Java9Parser";
+import { ClassType_lf_classOrInterfaceTypeContext } from "./Java9Parser";
+import { ClassType_lfno_classOrInterfaceTypeContext } from "./Java9Parser";
+import { InterfaceTypeContext } from "./Java9Parser";
+import { InterfaceType_lf_classOrInterfaceTypeContext } from "./Java9Parser";
+import { InterfaceType_lfno_classOrInterfaceTypeContext } from "./Java9Parser";
+import { TypeVariableContext } from "./Java9Parser";
+import { ArrayTypeContext } from "./Java9Parser";
+import { DimsContext } from "./Java9Parser";
+import { TypeParameterContext } from "./Java9Parser";
+import { TypeParameterModifierContext } from "./Java9Parser";
+import { TypeBoundContext } from "./Java9Parser";
+import { AdditionalBoundContext } from "./Java9Parser";
+import { TypeArgumentsContext } from "./Java9Parser";
+import { TypeArgumentListContext } from "./Java9Parser";
+import { TypeArgumentContext } from "./Java9Parser";
+import { WildcardContext } from "./Java9Parser";
+import { WildcardBoundsContext } from "./Java9Parser";
+import { ModuleNameContext } from "./Java9Parser";
+import { PackageNameContext } from "./Java9Parser";
+import { TypeNameContext } from "./Java9Parser";
+import { PackageOrTypeNameContext } from "./Java9Parser";
+import { ExpressionNameContext } from "./Java9Parser";
+import { MethodNameContext } from "./Java9Parser";
+import { AmbiguousNameContext } from "./Java9Parser";
+import { CompilationUnitContext } from "./Java9Parser";
+import { OrdinaryCompilationContext } from "./Java9Parser";
+import { ModularCompilationContext } from "./Java9Parser";
+import { PackageDeclarationContext } from "./Java9Parser";
+import { PackageModifierContext } from "./Java9Parser";
+import { ImportDeclarationContext } from "./Java9Parser";
+import { SingleTypeImportDeclarationContext } from "./Java9Parser";
+import { TypeImportOnDemandDeclarationContext } from "./Java9Parser";
+import { SingleStaticImportDeclarationContext } from "./Java9Parser";
+import { StaticImportOnDemandDeclarationContext } from "./Java9Parser";
+import { TypeDeclarationContext } from "./Java9Parser";
+import { ModuleDeclarationContext } from "./Java9Parser";
+import { ModuleDirectiveContext } from "./Java9Parser";
+import { RequiresModifierContext } from "./Java9Parser";
+import { ClassDeclarationContext } from "./Java9Parser";
+import { NormalClassDeclarationContext } from "./Java9Parser";
+import { ClassModifierContext } from "./Java9Parser";
+import { TypeParametersContext } from "./Java9Parser";
+import { TypeParameterListContext } from "./Java9Parser";
+import { SuperclassContext } from "./Java9Parser";
+import { SuperinterfacesContext } from "./Java9Parser";
+import { InterfaceTypeListContext } from "./Java9Parser";
+import { ClassBodyContext } from "./Java9Parser";
+import { ClassBodyDeclarationContext } from "./Java9Parser";
+import { ClassMemberDeclarationContext } from "./Java9Parser";
+import { FieldDeclarationContext } from "./Java9Parser";
+import { FieldModifierContext } from "./Java9Parser";
+import { VariableDeclaratorListContext } from "./Java9Parser";
+import { VariableDeclaratorContext } from "./Java9Parser";
+import { VariableDeclaratorIdContext } from "./Java9Parser";
+import { VariableInitializerContext } from "./Java9Parser";
+import { UnannTypeContext } from "./Java9Parser";
+import { UnannPrimitiveTypeContext } from "./Java9Parser";
+import { UnannReferenceTypeContext } from "./Java9Parser";
+import { UnannClassOrInterfaceTypeContext } from "./Java9Parser";
+import { UnannClassTypeContext } from "./Java9Parser";
+import { UnannClassType_lf_unannClassOrInterfaceTypeContext } from "./Java9Parser";
+import { UnannClassType_lfno_unannClassOrInterfaceTypeContext } from "./Java9Parser";
+import { UnannInterfaceTypeContext } from "./Java9Parser";
+import { UnannInterfaceType_lf_unannClassOrInterfaceTypeContext } from "./Java9Parser";
+import { UnannInterfaceType_lfno_unannClassOrInterfaceTypeContext } from "./Java9Parser";
+import { UnannTypeVariableContext } from "./Java9Parser";
+import { UnannArrayTypeContext } from "./Java9Parser";
+import { MethodDeclarationContext } from "./Java9Parser";
+import { MethodModifierContext } from "./Java9Parser";
+import { MethodHeaderContext } from "./Java9Parser";
+import { ResultContext } from "./Java9Parser";
+import { MethodDeclaratorContext } from "./Java9Parser";
+import { FormalParameterListContext } from "./Java9Parser";
+import { FormalParametersContext } from "./Java9Parser";
+import { FormalParameterContext } from "./Java9Parser";
+import { VariableModifierContext } from "./Java9Parser";
+import { LastFormalParameterContext } from "./Java9Parser";
+import { ReceiverParameterContext } from "./Java9Parser";
+import { Throws_Context } from "./Java9Parser";
+import { ExceptionTypeListContext } from "./Java9Parser";
+import { ExceptionTypeContext } from "./Java9Parser";
+import { MethodBodyContext } from "./Java9Parser";
+import { InstanceInitializerContext } from "./Java9Parser";
+import { StaticInitializerContext } from "./Java9Parser";
+import { ConstructorDeclarationContext } from "./Java9Parser";
+import { ConstructorModifierContext } from "./Java9Parser";
+import { ConstructorDeclaratorContext } from "./Java9Parser";
+import { SimpleTypeNameContext } from "./Java9Parser";
+import { ConstructorBodyContext } from "./Java9Parser";
+import { ExplicitConstructorInvocationContext } from "./Java9Parser";
+import { EnumDeclarationContext } from "./Java9Parser";
+import { EnumBodyContext } from "./Java9Parser";
+import { EnumConstantListContext } from "./Java9Parser";
+import { EnumConstantContext } from "./Java9Parser";
+import { EnumConstantModifierContext } from "./Java9Parser";
+import { EnumBodyDeclarationsContext } from "./Java9Parser";
+import { InterfaceDeclarationContext } from "./Java9Parser";
+import { NormalInterfaceDeclarationContext } from "./Java9Parser";
+import { InterfaceModifierContext } from "./Java9Parser";
+import { ExtendsInterfacesContext } from "./Java9Parser";
+import { InterfaceBodyContext } from "./Java9Parser";
+import { InterfaceMemberDeclarationContext } from "./Java9Parser";
+import { ConstantDeclarationContext } from "./Java9Parser";
+import { ConstantModifierContext } from "./Java9Parser";
+import { InterfaceMethodDeclarationContext } from "./Java9Parser";
+import { InterfaceMethodModifierContext } from "./Java9Parser";
+import { AnnotationTypeDeclarationContext } from "./Java9Parser";
+import { AnnotationTypeBodyContext } from "./Java9Parser";
+import { AnnotationTypeMemberDeclarationContext } from "./Java9Parser";
+import { AnnotationTypeElementDeclarationContext } from "./Java9Parser";
+import { AnnotationTypeElementModifierContext } from "./Java9Parser";
+import { DefaultValueContext } from "./Java9Parser";
+import { AnnotationContext } from "./Java9Parser";
+import { NormalAnnotationContext } from "./Java9Parser";
+import { ElementValuePairListContext } from "./Java9Parser";
+import { ElementValuePairContext } from "./Java9Parser";
+import { ElementValueContext } from "./Java9Parser";
+import { ElementValueArrayInitializerContext } from "./Java9Parser";
+import { ElementValueListContext } from "./Java9Parser";
+import { MarkerAnnotationContext } from "./Java9Parser";
+import { SingleElementAnnotationContext } from "./Java9Parser";
+import { ArrayInitializerContext } from "./Java9Parser";
+import { VariableInitializerListContext } from "./Java9Parser";
+import { BlockContext } from "./Java9Parser";
+import { BlockStatementsContext } from "./Java9Parser";
+import { BlockStatementContext } from "./Java9Parser";
+import { LocalVariableDeclarationStatementContext } from "./Java9Parser";
+import { LocalVariableDeclarationContext } from "./Java9Parser";
+import { StatementContext } from "./Java9Parser";
+import { StatementNoShortIfContext } from "./Java9Parser";
+import { StatementWithoutTrailingSubstatementContext } from "./Java9Parser";
+import { EmptyStatementContext } from "./Java9Parser";
+import { LabeledStatementContext } from "./Java9Parser";
+import { LabeledStatementNoShortIfContext } from "./Java9Parser";
+import { ExpressionStatementContext } from "./Java9Parser";
+import { StatementExpressionContext } from "./Java9Parser";
+import { IfThenStatementContext } from "./Java9Parser";
+import { IfThenElseStatementContext } from "./Java9Parser";
+import { IfThenElseStatementNoShortIfContext } from "./Java9Parser";
+import { AssertStatementContext } from "./Java9Parser";
+import { SwitchStatementContext } from "./Java9Parser";
+import { SwitchBlockContext } from "./Java9Parser";
+import { SwitchBlockStatementGroupContext } from "./Java9Parser";
+import { SwitchLabelsContext } from "./Java9Parser";
+import { SwitchLabelContext } from "./Java9Parser";
+import { EnumConstantNameContext } from "./Java9Parser";
+import { WhileStatementContext } from "./Java9Parser";
+import { WhileStatementNoShortIfContext } from "./Java9Parser";
+import { DoStatementContext } from "./Java9Parser";
+import { ForStatementContext } from "./Java9Parser";
+import { ForStatementNoShortIfContext } from "./Java9Parser";
+import { BasicForStatementContext } from "./Java9Parser";
+import { BasicForStatementNoShortIfContext } from "./Java9Parser";
+import { ForInitContext } from "./Java9Parser";
+import { ForUpdateContext } from "./Java9Parser";
+import { StatementExpressionListContext } from "./Java9Parser";
+import { EnhancedForStatementContext } from "./Java9Parser";
+import { EnhancedForStatementNoShortIfContext } from "./Java9Parser";
+import { BreakStatementContext } from "./Java9Parser";
+import { ContinueStatementContext } from "./Java9Parser";
+import { ReturnStatementContext } from "./Java9Parser";
+import { ThrowStatementContext } from "./Java9Parser";
+import { SynchronizedStatementContext } from "./Java9Parser";
+import { TryStatementContext } from "./Java9Parser";
+import { CatchesContext } from "./Java9Parser";
+import { CatchClauseContext } from "./Java9Parser";
+import { CatchFormalParameterContext } from "./Java9Parser";
+import { CatchTypeContext } from "./Java9Parser";
+import { Finally_Context } from "./Java9Parser";
+import { TryWithResourcesStatementContext } from "./Java9Parser";
+import { ResourceSpecificationContext } from "./Java9Parser";
+import { ResourceListContext } from "./Java9Parser";
+import { ResourceContext } from "./Java9Parser";
+import { VariableAccessContext } from "./Java9Parser";
+import { PrimaryContext } from "./Java9Parser";
+import { PrimaryNoNewArrayContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lf_arrayAccessContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lfno_arrayAccessContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lf_primaryContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lf_primary_lf_arrayAccess_lf_primaryContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lf_primary_lfno_arrayAccess_lf_primaryContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lfno_primaryContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lfno_primary_lf_arrayAccess_lfno_primaryContext } from "./Java9Parser";
+import { PrimaryNoNewArray_lfno_primary_lfno_arrayAccess_lfno_primaryContext } from "./Java9Parser";
+import { ClassLiteralContext } from "./Java9Parser";
+import { ClassInstanceCreationExpressionContext } from "./Java9Parser";
+import { ClassInstanceCreationExpression_lf_primaryContext } from "./Java9Parser";
+import { ClassInstanceCreationExpression_lfno_primaryContext } from "./Java9Parser";
+import { TypeArgumentsOrDiamondContext } from "./Java9Parser";
+import { FieldAccessContext } from "./Java9Parser";
+import { FieldAccess_lf_primaryContext } from "./Java9Parser";
+import { FieldAccess_lfno_primaryContext } from "./Java9Parser";
+import { ArrayAccessContext } from "./Java9Parser";
+import { ArrayAccess_lf_primaryContext } from "./Java9Parser";
+import { ArrayAccess_lfno_primaryContext } from "./Java9Parser";
+import { MethodInvocationContext } from "./Java9Parser";
+import { MethodInvocation_lf_primaryContext } from "./Java9Parser";
+import { MethodInvocation_lfno_primaryContext } from "./Java9Parser";
+import { ArgumentListContext } from "./Java9Parser";
+import { MethodReferenceContext } from "./Java9Parser";
+import { MethodReference_lf_primaryContext } from "./Java9Parser";
+import { MethodReference_lfno_primaryContext } from "./Java9Parser";
+import { ArrayCreationExpressionContext } from "./Java9Parser";
+import { DimExprsContext } from "./Java9Parser";
+import { DimExprContext } from "./Java9Parser";
+import { ConstantExpressionContext } from "./Java9Parser";
+import { ExpressionContext } from "./Java9Parser";
+import { LambdaExpressionContext } from "./Java9Parser";
+import { LambdaParametersContext } from "./Java9Parser";
+import { InferredFormalParameterListContext } from "./Java9Parser";
+import { LambdaBodyContext } from "./Java9Parser";
+import { AssignmentExpressionContext } from "./Java9Parser";
+import { AssignmentContext } from "./Java9Parser";
+import { LeftHandSideContext } from "./Java9Parser";
+import { AssignmentOperatorContext } from "./Java9Parser";
+import { ConditionalExpressionContext } from "./Java9Parser";
+import { ConditionalOrExpressionContext } from "./Java9Parser";
+import { ConditionalAndExpressionContext } from "./Java9Parser";
+import { InclusiveOrExpressionContext } from "./Java9Parser";
+import { ExclusiveOrExpressionContext } from "./Java9Parser";
+import { AndExpressionContext } from "./Java9Parser";
+import { EqualityExpressionContext } from "./Java9Parser";
+import { RelationalExpressionContext } from "./Java9Parser";
+import { ShiftExpressionContext } from "./Java9Parser";
+import { AdditiveExpressionContext } from "./Java9Parser";
+import { MultiplicativeExpressionContext } from "./Java9Parser";
+import { UnaryExpressionContext } from "./Java9Parser";
+import { PreIncrementExpressionContext } from "./Java9Parser";
+import { PreDecrementExpressionContext } from "./Java9Parser";
+import { UnaryExpressionNotPlusMinusContext } from "./Java9Parser";
+import { PostfixExpressionContext } from "./Java9Parser";
+import { PostIncrementExpressionContext } from "./Java9Parser";
+import { PostIncrementExpression_lf_postfixExpressionContext } from "./Java9Parser";
+import { PostDecrementExpressionContext } from "./Java9Parser";
+import { PostDecrementExpression_lf_postfixExpressionContext } from "./Java9Parser";
+import { CastExpressionContext } from "./Java9Parser";
+import { IdentifierContext } from "./Java9Parser";
 
 
 /**
